+++ src/modules/media/media.module.ts
@@ -4,9 +4,10 @@ import { MediaService } from './media.service';
 import { StorageModule } from '../../shared/storage/storage.module';
 import { PrismaService } from '../../prisma/prisma.service';
 import { UploadsModule } from '../uploads/uploads.module';
+import { MediaJobsModule } from '../media-jobs/media-jobs.module';
 
 @Module({
-  imports: [StorageModule, UploadsModule],
+  imports: [StorageModule, UploadsModule, MediaJobsModule],
   controllers: [MediaController],
   providers: [PrismaService, MediaService],
 })